Eliminate unnecessary dependency on bfd.h from sepdebugcrcfix
I think the reliance on bfd.h was a mistake. The code was lifted from
bfd, but should be totally independent (it just calculates a CRC).
Fix the type to be a normal size_t and include sys/stat.h (which was
included through bfd.h) to get the definitions of stat and chmod.

(cherry picked from commit 245b5a3b4b6d616adf47361137987e90f8dab22c)
Signed-off-by: David Abdurachmanov <david.abdurachmanov@gmail.com>